% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presents.R
\name{theme_carly_presents}
\alias{theme_carly_presents}
\title{Presentation Theme}
\usage{
theme_carly_presents(base_size = 16, base_family = "",
  base_line_size = base_size/22, base_rect_size = base_size/22)
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\description{
I designed this theme for presentation. It features sharp constrast and larger fonts.
It excludes details, like gridlines, that might not display clearly in a low-res format.
}
\details{
\describe{

\item{`theme_carly_presents`}{
The signature ggplot2 theme with a grey background and white gridlines,
designed to put the data forward yet make comparisons easy.}
}
}
\examples{
library(ggplot2)
n <- 100
x <- rnorm(n)
y <- rnorm(n, x)
p <- qplot(x, y)
p + theme_carly_presents()
}
